+++ api/src/main/java/ca/bc/gov/educ/api/batchgraduation/model/GraduationStudentRecord.java
@@ -21,6 +21,7 @@ public class GraduationStudentRecord extends BaseModel {
     private String honoursStanding;
     private String recalculateGradStatus;   
     private String schoolOfRecord;
+    private UUID schoolOfRecordId;
     private String schoolName;
     private String studentGrade;	
     private String studentStatus;
